+++ components/common/ProfileImage.tsx
@@ -10,7 +10,7 @@ export default function ProfileImage({ name, src, size }: ProfileImageProps) {
   const divSize = size === 'sm' ? 'w-26pxr h-26pxr' : 'w-38pxr h-38pxr';
   return (
     <div
-      className={`rounded-full flex-center bg-green relative overflow-hidden ${divSize}`}
+      className={`rounded-full flex-center bg-green border-2 border-white relative overflow-hidden ${divSize}`}
     >
       {src ? (
         <Image src={src} alt="프로필 이미지" fill />
