+++ components/dashboard/header/DashboardHeader.tsx
@@ -1,7 +1,6 @@
 import { useRouter } from 'next/router';
-import Header from './Header';
+import { Header, HeaderMembers, HeaderButton } from '@/components';
 import { DashboardHeaderProps } from './Header';
-import HeaderButton from './HeaderButton';
 import InviteButton from './InviteButton';
 import settingIcon from '@/public/icons/setting-icon.svg';
 
@@ -10,6 +9,7 @@ export default function DashboardHeader({ dashboard }: DashboardHeaderProps) {
   const handleSettingButtonClick = () => {
     router.push(`${dashboard.id}/edit`);
   };
+
   return (
     <Header dashboard={dashboard}>
       <div className="flex items-center desktop:gap-32pxr gap-24pxr mobile:gap-12pxr">
@@ -23,7 +23,7 @@ export default function DashboardHeader({ dashboard }: DashboardHeaderProps) {
           </HeaderButton>
           <InviteButton />
         </div>
-        <div className="desktop:w-158pxr h-34pxr w-98pxr mobile:w-82pxr bg-pink" />
+        <HeaderMembers dashboardId={dashboard.id} />
         <div className="w-1pxr h-38pxr desktop:mr-32pxr mr-24pxr mobile:mr-12pxr bg-gray30" />
       </div>
     </Header>
