+++ components/index.ts
@@ -6,6 +6,14 @@ export { default as ProfileImage } from './common/ProfileImage';
 export { default as Profile } from './common/Profile';
 export { default as PasswordInput } from './common/PasswordInput';
 export { default as Toast } from './common/Toast';
+export { default as Portal } from './modal/Portal';
+export { default as ModalTitle } from './modal/ModalTitle';
+export { default as ModalButton } from './modal/ModalButton';
+export { default as ImagePick } from './modal/ImagePick';
+export { default as SelectDate } from './modal/SelectDate';
+export { default as Modal } from './modal/Modal';
+export { default as InviteModal } from './modal/Invite/InviteModal';
+export { default as CreateDashboardModal } from './modal/create-dashboard/CreateDashboardModal';
 export { default as InviteListTable } from './boardEdit/InviteListTable';
 export { default as MembersTable } from './boardEdit/MembersTable';
 export { default as NameEditForm } from './boardEdit/NameEditForm';
@@ -17,6 +25,7 @@ export { default as SignUpForm } from './sign/form/SignUpForm';
 export { default as MypageProfile } from './mypage/MypageProfile';
 export { default as PasswordModify } from './mypage/PasswordModify';
 export { default as HeaderButton } from './dashboard/header/HeaderButton';
+export { default as HeaderMembers } from './dashboard/header/HeaderMembers';
 export { default as Header } from './dashboard/header/Header';
 export { default as DashboardHeader } from './dashboard/header/DashboardHeader';
 export { default as CreateDashboardButton } from './dashboard/sidebar/CreateDashboardButton';
